chore: Move tooltips registry to internal folder
