+++ src/internal/components/tooltip/registry.ts
